+++ common/src/main/java/net/elemental_wizards_rpg/client/ElementalClient.java
@@ -1,39 +1,22 @@
 package net.elemental_wizards_rpg.client;
 
-import mod.azure.azurelibarmor.common.render.armor.AzArmorRenderer;
-import mod.azure.azurelibarmor.common.render.armor.AzArmorRendererRegistry;
+import net.rpg_foundation.armor_api.client.ArmorRenderers;
+import net.rpg_foundation.armor_api.client.GeoArmorRenderer;
 import net.elemental_wizards_rpg.client.armor.ElementalRobeRenderer;
 import net.elemental_wizards_rpg.client.effect.*;
-import net.elemental_wizards_rpg.client.entity.WhirlwindRenderer;
-import net.elemental_wizards_rpg.client.entity.TerraStoneRenderer;
-import net.elemental_wizards_rpg.client.entity.earth_golem.EarthGolemEntityRenderer;
-import net.elemental_wizards_rpg.client.entity.earth_golem.EarthGolemSpikeEntityRenderer;
-import net.elemental_wizards_rpg.client.entity.TidalWaveEntityRenderer;
-import net.elemental_wizards_rpg.client.entity.HealingRainCloudEntityRenderer;
-import net.elemental_wizards_rpg.client.entity.EarthquakeEntityRenderer;
 import net.elemental_wizards_rpg.effect.ElementalEffects;
-import net.elemental_wizards_rpg.entity.WhirlwindEntity;
-import net.elemental_wizards_rpg.entity.TerraStoneEntity;
-import net.elemental_wizards_rpg.entity.EarthGolemEntity;
-import net.elemental_wizards_rpg.entity.EarthGolemSpikeEntity;
-import net.elemental_wizards_rpg.entity.TidalWaveEntity;
-import net.elemental_wizards_rpg.entity.HealingRainCloudEntity;
-import net.elemental_wizards_rpg.entity.EarthquakeEntity;
 import net.elemental_wizards_rpg.item.armor.Armors;
 import net.elemental_wizards_rpg.particle.ModParticles;
 import net.elemental_wizards_rpg.spell.ElementalWizardSpells;
-import net.elemental_wizards_rpg.client.entity.earth_golem.EarthGolemEntityModel;
-import net.fabricmc.fabric.api.client.particle.v1.ParticleFactoryRegistry;
-import net.fabricmc.fabric.api.client.rendering.v1.EntityModelLayerRegistry;
-import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
-import net.fabricmc.fabric.api.client.rendering.v1.LivingEntityFeatureRendererRegistrationCallback;
-import net.minecraft.client.render.entity.PlayerEntityRenderer;
+import net.minecraft.client.particle.ParticleFactory;
+import net.minecraft.client.particle.SpriteProvider;
+import net.minecraft.particle.ParticleType;
 import net.spell_engine.api.effect.CustomModelStatusEffect;
 import net.spell_engine.api.effect.CustomParticleStatusEffect;
 import net.spell_engine.client.particle.SpellParticle;
 import net.spell_engine.rpg_series.item.Armor;
 
-import java.util.function.Supplier;
+import java.util.function.Function;
 
 import static net.elemental_wizards_rpg.compat.CompatLoadingCheck.armoryLoadCheck;
 
@@ -43,18 +26,18 @@ public class ElementalClient{
     public static void init(){
         ElementalWizardSpells.registerTooltipTokens();
 
-        registerArmorRenderer(Armors.elementalArmor.armorSet(), ElementalRobeRenderer::elemental);
-        registerArmorRenderer(Armors.kelpArmor.armorSet(), ElementalRobeRenderer::kelp);
-        registerArmorRenderer(Armors.dripstoneArmor.armorSet(), ElementalRobeRenderer::dripstone);
-        registerArmorRenderer(Armors.windArmor.armorSet(), ElementalRobeRenderer::wind);
-        registerArmorRenderer(Armors.netheriteKelpNetheriteArmor.armorSet(), ElementalRobeRenderer::netherite_kelp);
-        registerArmorRenderer(Armors.netheriteDripstoneArmor.armorSet(), ElementalRobeRenderer::netherite_dripstone);
-        registerArmorRenderer(Armors.netheriteWindArmor.armorSet(), ElementalRobeRenderer::netherite_wind);
+        registerArmorRenderer(Armors.elementalArmor.armorSet(), ElementalRobeRenderer.elemental());
+        registerArmorRenderer(Armors.kelpArmor.armorSet(), ElementalRobeRenderer.kelp());
+        registerArmorRenderer(Armors.dripstoneArmor.armorSet(), ElementalRobeRenderer.dripstone());
+        registerArmorRenderer(Armors.windArmor.armorSet(), ElementalRobeRenderer.wind());
+        registerArmorRenderer(Armors.netheriteKelpNetheriteArmor.armorSet(), ElementalRobeRenderer.netherite_kelp());
+        registerArmorRenderer(Armors.netheriteDripstoneArmor.armorSet(), ElementalRobeRenderer.netherite_dripstone());
+        registerArmorRenderer(Armors.netheriteWindArmor.armorSet(), ElementalRobeRenderer.netherite_wind());
 
         if (armoryLoadCheck()) {
-            registerArmorRenderer(Armors.hurricaneArmorSet.armorSet(), ElementalRobeRenderer::hurricane);
-            registerArmorRenderer(Armors.mountainArmorSet.armorSet(), ElementalRobeRenderer::mountain);
-            registerArmorRenderer(Armors.oceanArmorSet.armorSet(), ElementalRobeRenderer::ocean);
+            registerArmorRenderer(Armors.hurricaneArmorSet.armorSet(), ElementalRobeRenderer.hurricane());
+            registerArmorRenderer(Armors.mountainArmorSet.armorSet(), ElementalRobeRenderer.mountain());
+            registerArmorRenderer(Armors.oceanArmorSet.armorSet(), ElementalRobeRenderer.ocean());
         }
 
         CustomModelStatusEffect.register(ElementalEffects.BUBBLE_FOAM.effect, new BubbleFoamEffectRenderer());
@@ -63,33 +46,20 @@ public class ElementalClient{
         CustomParticleStatusEffect.register(ElementalEffects.BUBBLE_FOAM.effect, new BubbleFoamParticleSpawner());
         CustomParticleStatusEffect.register(ElementalEffects.STONE_FLESH.effect, new StoneFleshParticleSpawner());
 
-        LivingEntityFeatureRendererRegistrationCallback.EVENT.register((entityType, entityRenderer, registrationHelper, context) -> {
-            if (entityRenderer instanceof PlayerEntityRenderer playerRenderer) {
-                registrationHelper.register(new StoneFleshPlayerRenderLayer(playerRenderer));
-            }
-        });
-
         CustomModelStatusEffect.register(ElementalEffects.IMPALED.effect, new ImpaledRenderer());
-
-        EntityModelLayerRegistry.registerModelLayer(EarthGolemEntityModel.LAYER_LOCATION, EarthGolemEntityModel::createBodyLayer);
-
-        EntityRendererRegistry.register(WhirlwindEntity.ENTITY_TYPE, WhirlwindRenderer::new);
-        EntityRendererRegistry.register(TerraStoneEntity.ENTITY_TYPE, TerraStoneRenderer::new);
-        EntityRendererRegistry.register(EarthGolemEntity.ENTITY_TYPE, EarthGolemEntityRenderer::new);
-        EntityRendererRegistry.register(EarthGolemSpikeEntity.ENTITY_TYPE, EarthGolemSpikeEntityRenderer::new);
-        EntityRendererRegistry.register(TidalWaveEntity.ENTITY_TYPE, TidalWaveEntityRenderer::new);
-        EntityRendererRegistry.register(HealingRainCloudEntity.ENTITY_TYPE, HealingRainCloudEntityRenderer::new);
-        EntityRendererRegistry.register(EarthquakeEntity.ENTITY_TYPE, EarthquakeEntityRenderer::new);
-
     }
-    private static void registerArmorRenderer(Armor.Set set, Supplier<AzArmorRenderer> armorRendererSupplier) {
-        AzArmorRendererRegistry.register(armorRendererSupplier, set.head, set.chest, set.legs, set.feet);
+    private static void registerArmorRenderer(Armor.Set set, GeoArmorRenderer renderer) {
+        ArmorRenderers.register(renderer, set.head, set.chest, set.legs, set.feet);
     }
-    public static void registerParticleAppearances() {
-        ParticleFactoryRegistry registry = ParticleFactoryRegistry.getInstance();
 
+    @FunctionalInterface
+    public interface ParticleFactoryRegistrar {
+        void register(ParticleType type, Function<SpriteProvider, ParticleFactory> factory);
+    }
+
+    public static void registerParticleAppearances(ParticleFactoryRegistrar registrar) {
         for (var entry: ModParticles.entries()) {
-            registry.register(entry.type(), provider -> new SpellParticle.Factory(provider, entry));
+            registrar.register(entry.type(), provider -> new SpellParticle.Factory(provider, entry));
         }
     }
 }