+++ common/src/main/java/net/elemental_wizards_rpg/client/armor/ElementalRobeRenderer.java
@@ -1,50 +1,50 @@
 package net.elemental_wizards_rpg.client.armor;
 
-import mod.azure.azurelibarmor.common.render.armor.AzArmorRenderer;
-import mod.azure.azurelibarmor.common.render.armor.AzArmorRendererConfig;
 import net.minecraft.util.Identifier;
+import net.rpg_foundation.armor_api.client.GeoArmorRenderer;
 
 import static net.elemental_wizards_rpg.ElementalMod.MOD_ID;
 
-public class ElementalRobeRenderer extends AzArmorRenderer {
-    public static ElementalRobeRenderer elemental() {
-        return new ElementalRobeRenderer("wizard_t1", "elemental");
+public final class ElementalRobeRenderer {
+    private ElementalRobeRenderer() { }
+
+    public static GeoArmorRenderer elemental() {
+        return make("wizard_t1", "elemental");
     }
 
-    public static ElementalRobeRenderer kelp() {
-        return new ElementalRobeRenderer("wizard_t1", "kelp");
+    public static GeoArmorRenderer kelp() {
+        return make("wizard_t1", "kelp");
     }
-    public static ElementalRobeRenderer dripstone() {
-        return new ElementalRobeRenderer("wizard_t1", "dripstone");
+    public static GeoArmorRenderer dripstone() {
+        return make("wizard_t1", "dripstone");
     }
-    public static ElementalRobeRenderer wind() {
-        return new ElementalRobeRenderer("wizard_t1", "wind");
-    }
-
-    public static ElementalRobeRenderer netherite_kelp() {
-        return new ElementalRobeRenderer("wizard_t1", "netherite_kelp");
-    }
-    public static ElementalRobeRenderer netherite_dripstone() {
-        return new ElementalRobeRenderer("wizard_t1", "netherite_dripstone");
-    }
-    public static ElementalRobeRenderer netherite_wind() {
-        return new ElementalRobeRenderer("wizard_t1", "netherite_wind");
+    public static GeoArmorRenderer wind() {
+        return make("wizard_t1", "wind");
     }
 
-    public static ElementalRobeRenderer hurricane() {
-        return new ElementalRobeRenderer("hurricane", "hurricane");
+    public static GeoArmorRenderer netherite_kelp() {
+        return make("wizard_t1", "netherite_kelp");
     }
-    public static ElementalRobeRenderer mountain() {
-        return new ElementalRobeRenderer("mountain", "mountain");
+    public static GeoArmorRenderer netherite_dripstone() {
+        return make("wizard_t1", "netherite_dripstone");
     }
-    public static ElementalRobeRenderer ocean() {
-        return new ElementalRobeRenderer("ocean", "ocean");
+    public static GeoArmorRenderer netherite_wind() {
+        return make("wizard_t1", "netherite_wind");
     }
 
-    public ElementalRobeRenderer(String modelName, String textureName) {
-        super(AzArmorRendererConfig.builder(
+    public static GeoArmorRenderer hurricane() {
+        return make("hurricane", "hurricane");
+    }
+    public static GeoArmorRenderer mountain() {
+        return make("mountain", "mountain");
+    }
+    public static GeoArmorRenderer ocean() {
+        return make("ocean", "ocean");
+    }
+
+    private static GeoArmorRenderer make(String modelName, String textureName) {
+        return GeoArmorRenderer.of(
                 Identifier.of(MOD_ID, "geo/" + modelName + ".geo.json"),
-                Identifier.of(MOD_ID, "textures/armor/" + textureName + ".png")
-        ).build());
+                Identifier.of(MOD_ID, "textures/armor/" + textureName + ".png"));
     }
 }