+++ common/src/main/java/net/elemental_wizards_rpg/item/ElementalGroup.java
@@ -1,12 +1,12 @@
 package net.elemental_wizards_rpg.item;
 
 import net.elemental_wizards_rpg.ElementalMod;
-import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroupEntries;
-import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
+import net.elemental_wizards_rpg.item.armor.Armors;
 import net.minecraft.item.ItemGroup;
-import net.minecraft.item.ItemGroups;
+import net.minecraft.item.ItemStack;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.RegistryKey;
+import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 
 public class ElementalGroup {
@@ -14,11 +14,15 @@ public class ElementalGroup {
     public static RegistryKey<ItemGroup> ELEMENTAL_WIZARD_KEY = RegistryKey.of(Registries.ITEM_GROUP.getKey(),Identifier.of(ElementalMod.MOD_ID,"generic"));
     public static ItemGroup ELEMENTAL_WIZARD;
 
-    private static void addItemsToIngredientItemGroup(FabricItemGroupEntries entries){
+    public static void addItemsToIngredientItemGroup(ItemGroup.Entries entries){
         entries.add(ElementalItems.ELEMENTAL_ESSENCE);
     }
 
-    public static void registerItemGroups() {
-        ItemGroupEvents.modifyEntriesEvent(ItemGroups.INGREDIENTS).register(ElementalGroup::addItemsToIngredientItemGroup);
+    public static ItemStack icon() {
+        return new ItemStack(Armors.kelpArmor.armorSet().head);
+    }
+
+    public static Text displayName() {
+        return Text.translatable("itemGroup." + ElementalMod.MOD_ID + ".general");
     }
 }